Initial 0.7.0 update using neo4j-uploader and updated graph-data-generator package for uploading directly to a Neo4j database
