+++ Pipfile
@@ -7,6 +7,7 @@ name = "pypi"
 streamlit-player = "*"
 graph-data-generator = "*"
 streamlit-extras = "*"
+neo4j-uploader = "*"
 
 [dev-packages]
