+++ graph_data_generator_streamlit/ui/data_importer_ui.py
@@ -2,7 +2,7 @@ import streamlit as st
 import streamlit.components.v1 as components
 from streamlit_extras.colored_header import colored_header
 
-def data_importer_tab():
+def data_importer_ui():
     is_local = st.checkbox("Use HTTP", value=False, help="Select Use HTTP if connecting with a local Neo4j instance.")
 
     if is_local == True:
