+++ sample_optional_secrets.toml
@@ -1,0 +1,4 @@
+OPENAI_API_KEY=""
+NEO4J_URI=""
+NEO4J_USER="neo4j"
+NEO4J_PASSWORD=""
